Tested Spectral code on all 3 sites. Change TReceive to 1 byte. Reinstate some HAL i2c functions.

+++ src/esw/fw/science/Core/Inc/i2c_mux.hpp
@@ -28,12 +28,12 @@ namespace mrover {
     public:
     	I2CMux() = default;
 
-    	I2CMux(std::shared_ptr<SMBus<uint8_t, uint16_t>> i2c_bus, Pin reset_pin);
+    	I2CMux(std::shared_ptr<SMBus<uint8_t, uint8_t>> i2c_bus, Pin reset_pin);
 
         void set_channel(uint8_t channel);
 
     private:
-        std::shared_ptr<SMBus<uint8_t, uint16_t>> m_i2c_bus;
+        std::shared_ptr<SMBus<uint8_t, uint8_t>> m_i2c_bus;
         uint8_t current_channel = 0;
         constexpr static std::uint16_t MUX_7b_ADDRESS = 0x70;
         Pin m_reset_pin;
