+++ src/esw/fw/science/Core/Src/i2c_mux.cpp
@@ -2,7 +2,7 @@
 // Hardware PCA9546A
 namespace mrover {
 
-	I2CMux::I2CMux(std::shared_ptr<SMBus<uint8_t, uint16_t>> i2c_bus, Pin reset_pin)
+	I2CMux::I2CMux(std::shared_ptr<SMBus<uint8_t, uint8_t>> i2c_bus, Pin reset_pin)
     	: m_i2c_bus(i2c_bus), m_reset_pin(reset_pin) {
 		reset_pin.write(GPIO_PIN_SET);
 	}
